pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["kokoserver <owonikokoolaoluwa@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.67.0"
uvicorn = "^0.18.2"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.2"}
pydantic = {extras = ["email", "dotenv"], version = "^1.9.2"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
Jinja2 = "^3.1.2"
python-multipart = "^0.0.5"
python-slugify = "^6.1.2"


[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-cov = "^2.10.1"
autoflake = "^1.4"
flake8 = "^3.8.4"
mypy = "^0.790"
isort = "^5.0"
pre-commit = "^2.8.2"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
